Integrated Firebase
added all firebase components from database to authentication to the project. Will need more work later on.
