+++ frontend/remindit-frontend/src/Front End Components/loginform.js
@@ -55,6 +55,7 @@ const SpanText = styled('span')(({ theme }) => ({
     }
 }))
 
+
 const LoginForm = () => {
 
     return (
